- removed need for int() conversion now that x << s has correct type - updated golden.out
R=r
DELTA=9  (0 added, 8 deleted, 1 changed)
OCL=14657
CL=14657

+++ test/hashmap.go
@@ -47,9 +47,7 @@ type HashMap struct {
 
 
 func (m *HashMap) capacity() uint32 {
-	// TODO we need to figure out how to determine the type of
-	// a shifted 'untyped' int so we can get rid of the conversion
-	return uint32(1) << m.log2_capacity_;
+	return 1 << m.log2_capacity_;
 }
